move axes options out of tabs file
